zobrist hashing and game status now in Game
