+++ src/types.ts
@@ -2,6 +2,47 @@
  * Shared types and constants used across chess.js modules
  */
 
+// Zobrist hashing utilities
+const MASK64 = 0xffffffffffffffffn
+
+function rotl(x: bigint, k: bigint): bigint {
+  return ((x << k) | (x >> (64n - k))) & 0xffffffffffffffffn
+}
+
+function wrappingMul(x: bigint, y: bigint) {
+  return (x * y) & MASK64
+}
+
+// xoroshiro128**
+export function xoroshiro128(state: bigint) {
+  return function () {
+    let s0 = BigInt(state & MASK64)
+    let s1 = BigInt((state >> 64n) & MASK64)
+
+    const result = wrappingMul(rotl(wrappingMul(s0, 5n), 7n), 9n)
+
+    s1 ^= s0
+    s0 = (rotl(s0, 24n) ^ s1 ^ (s1 << 16n)) & MASK64
+    s1 = rotl(s1, 37n)
+
+    state = (s1 << 64n) | s0
+
+    return result
+  }
+}
+
+const rand = xoroshiro128(0xa187eb39cdcaed8f31c4b365b102e01en)
+
+export const PIECE_KEYS = Array.from({ length: 2 }, () =>
+  Array.from({ length: 6 }, () => Array.from({ length: 128 }, () => rand())),
+)
+
+export const EP_KEYS = Array.from({ length: 8 }, () => rand())
+
+export const CASTLING_KEYS = Array.from({ length: 16 }, () => rand())
+
+export const SIDE_KEY = rand()
+
 export const WHITE = 'w'
 export const BLACK = 'b'
